table 선택 hook 분리
