+++ src/pages/products/product/product-page.tsx
@@ -1,6 +1,6 @@
 import FilterTabs from '../../../features/products/product/filter/filter-tabs'
 import FilterCategory from '../../../features/products/product/filter/filter-category'
-import ResultTable from '../../../features/products/product/product-list/result-table'
+import ResultTable from '../../../features/products/product/product-list/product-list-table'
 
 import { cn } from '@/shared/lib/utils'
